Move shortcuts implementation to the private module.
